modified functionspaces to uint32 instead of uint64. modified a few other things to remove the lookup table in the evaluator.

+++ lib-date.go
@@ -15,7 +15,7 @@ func buildDateLib() {
                                 "time_dow","time_dom","time_month","time_year" }
 
 	slhelp["date"] = LibHelp{in: "integer", out: "string", action: "Returns a date/time string. The parsed timestamp is either the current date/time or the optional [#i1]integer[#i0]. RFC3339 format."}
-	stdlib["date"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+	stdlib["date"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
 		var t time.Time
 		if len(args) == 1 {
 			when, invalid := GetAsInt64(args[0])
@@ -31,17 +31,17 @@ func buildDateLib() {
 	}
 
 	slhelp["date_human"] = LibHelp{in: "", out: "string", action: "Returns the current date and time in a readable format (RFC822Z)"}
-	stdlib["date_human"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+	stdlib["date_human"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         return time.Now().Format(time.RFC822Z),nil
     }
 
 	slhelp["epoch_time"] = LibHelp{in: "", out: "integer", action: "Returns the current epoch (Unix) time in seconds."}
-	stdlib["epoch_time"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+	stdlib["epoch_time"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
 		return int(time.Now().Unix()), err
 	}
 
 	slhelp["time_hours"] = LibHelp{in: "epochnano", out: "integer", action: "Returns the current hour of the day (no arguments) or the hour of the day specified by the epoch nano time in [#i1]epochnano[#i0]."}
-	stdlib["time_hours"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+	stdlib["time_hours"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
 		var t time.Time
 		if len(args) == 1 {
 			when, invalid := GetAsInt(args[0])
@@ -56,7 +56,7 @@ func buildDateLib() {
 	}
 
 	slhelp["time_minutes"] = LibHelp{in: "epochnano", out: "integer", action: "Returns the current minute of the day (no arguments) or the minute of the day specified by the epoch nano time in [#i1]epochnano[#i0]."}
-	stdlib["time_minutes"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+	stdlib["time_minutes"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
 		var t time.Time
 		if len(args) == 1 {
 			when, invalid := GetAsInt(args[0])
@@ -71,7 +71,7 @@ func buildDateLib() {
 	}
 
 	slhelp["time_seconds"] = LibHelp{in: "epochnano", out: "integer", action: "Returns the current second of the day (no arguments) or the second of the day specified by the epoch nano time in [#i1]epochnano[#i0]."}
-	stdlib["time_seconds"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+	stdlib["time_seconds"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
 		var t time.Time
 		if len(args) == 1 {
 			when, invalid := GetAsInt(args[0])
@@ -86,7 +86,7 @@ func buildDateLib() {
 	}
 
 	slhelp["time_nanos"] = LibHelp{in: "epochnano", out: "integer", action: "Returns the offset within the current second, in nanoseconds, (no arguments) or the offset within the epoch nano time specified in [#i1]epochnano[#i0]."}
-	stdlib["time_nanos"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+	stdlib["time_nanos"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
 		var t time.Time
 		if len(args) == 1 {
 			when, invalid := GetAsInt(args[0])
@@ -101,7 +101,7 @@ func buildDateLib() {
 	}
 
 	slhelp["time_dow"] = LibHelp{in: "epochnano", out: "string", action: "Returns the current day of the week (no arguments) or the day of the epoch nano time specified in [#i1]epochnano[#i0]."}
-	stdlib["time_dow"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+	stdlib["time_dow"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
 		var t time.Time
 		if len(args) == 1 {
 			when, invalid := GetAsInt(args[0])
@@ -116,7 +116,7 @@ func buildDateLib() {
 	}
 
 	slhelp["time_dom"] = LibHelp{in: "epochnano", out: "integer", action: "Returns the current day of the month (no arguments) or the day of the epoch nano time specified in [#i1]epochnano[#i0]."}
-	stdlib["time_dom"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+	stdlib["time_dom"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
 		var t time.Time
 		if len(args) == 1 {
 			when, invalid := GetAsInt(args[0])
@@ -131,7 +131,7 @@ func buildDateLib() {
 	}
 
 	slhelp["time_month"] = LibHelp{in: "epochnano", out: "integer", action: "Returns the current month of the year (no arguments) or the month of the epoch nano time specified in [#i1]epochnano[#i0]."}
-	stdlib["time_month"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+	stdlib["time_month"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
 		var t time.Time
 		if len(args) == 1 {
 			when, invalid := GetAsInt(args[0])
@@ -146,7 +146,7 @@ func buildDateLib() {
 	}
 
 	slhelp["time_year"] = LibHelp{in: "epochnano", out: "integer", action: "Returns the current year (no arguments) or the year of the epoch nano time specified in [#i1]epochnano[#i0]."}
-	stdlib["time_year"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+	stdlib["time_year"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
 		var t time.Time
 		if len(args) == 1 {
 			when, invalid := GetAsInt(args[0])
@@ -161,12 +161,12 @@ func buildDateLib() {
 	}
 
 	slhelp["epoch_nano_time"] = LibHelp{in: "", out: "integer", action: "Returns the current epoch (Unix) time in nano-seconds."}
-	stdlib["epoch_nano_time"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+	stdlib["epoch_nano_time"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
 		return int(time.Now().UnixNano()), err
 	}
 
 	slhelp["time_diff"] = LibHelp{in: "te,ts", out: "integer", action: "Returns difference in micro-seconds between two nano-epoch dates."}
-	stdlib["time_diff"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+	stdlib["time_diff"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
 
 		if len(args) != 2 {
 			return 0, errors.New("Bad args (count) for time_diff()")
